[NX-429] Move ReferenceXO back to core ui plugin
