+++ plugins/rapture/nexus-coreui-plugin/src/main/java/org/sonatype/nexus/coreui/RoleComponent.groovy
@@ -20,7 +20,6 @@ import org.apache.shiro.authz.annotation.RequiresAuthentication
 import org.apache.shiro.authz.annotation.RequiresPermissions
 import org.sonatype.nexus.extdirect.DirectComponent
 import org.sonatype.nexus.extdirect.DirectComponentSupport
-import org.sonatype.nexus.rapture.ReferenceXO
 import org.sonatype.nexus.validation.Create
 import org.sonatype.nexus.validation.Update
 import org.sonatype.security.SecuritySystem
